fixup! TF-4449 Add drive_attachment presentation layer and composer integration
